linshare-chat-client/requirements.txt

chainlit
mcp
fastmcp
httpx
python-dotenv
langchain
langchain-google-genai
google-generativeai
langchain-openai
langchain-groq
